unbreak two examples
missing two casts of float64 to int and the initial call to queue.get
must not be blocking so push a silence initial sample.

+++ examples/spectrogram.py
@@ -55,7 +55,7 @@ try:
 
     delta_f = (high - low) / (args.columns - 1)
     fftsize = np.ceil(samplerate / delta_f).astype(int)
-    low_bin = np.floor(low / delta_f)
+    low_bin = np.floor(low / delta_f).astype(int)
 
     cumulated_status = sd.CallbackFlags()
